Moving potentially colliding classes to active_mappers namespace
